+++ RING/func/Build.py
@@ -1,116 +1,7 @@
-from zeep import Client
-from zeep.cache import SqliteCache
-from zeep.transports import Transport
 from zeep.exceptions import Fault
-from zeep.plugins import HistoryPlugin
-from requests import Session
-from requests.auth import HTTPBasicAuth
-from urllib3 import disable_warnings
-from urllib3.exceptions import InsecureRequestWarning
-from lxml import etree
-import os
+import RING.conf as Config
 
-#
-# Config
-#
-
-RegionNamesList = ["SBC","BROADCAST"]
-WSDL = r'\axlsqltoolkit\schema\current\AXLAPI.wsdl'
-
-
-#
-# Variables
-#
-WSDL = f"file:\\\{os.path.dirname(os.path.realpath(__file__))}{WSDL}"
-
-#
-# Classes
-#
-class AxlConnection:
-    def __init__(self, wsdlpath):
-        self.wsdl = wsdlpath
-        self.host = '10.10.20.1'
-        self.binding = "{http://www.cisco.com/AXLAPIService/}AXLAPIBinding"
-        self.username = 'administrator'
-        self.password = 'ciscopsdt'
-        self.Service = None
-        self.__State = False
-
-    def ConnectionState(self):
-        return self.__State
-
-    def Open(self, timeout=20):
-        try:
-            disable_warnings(InsecureRequestWarning) # Disable warning output due to invalid certificate
-
-            session = Session()
-            session.verify = False #don't do this in production
-            session.auth = HTTPBasicAuth(self.username, self.password)
-
-            location = 'https://{host}:8443/axl/'.format(host=self.host)
-            transport = Transport(cache=SqliteCache(), session=session, timeout=timeout)
-            history = HistoryPlugin()
-            client = Client(wsdl=self.wsdl, transport=transport, plugins=[history])
-
-            self.Service = client.create_service(self.binding, location)
-            self.__State = True
-            return True
-
-        except Fault as err:
-            print (f"{err}")
-            return False
-
-    def __del__(self):
-        pass
-
-class Site:
-    def __init__(self, SiteCode):
-        self.SiteCode = SiteCode
-        self.ClusterNumber = 0
-        self.AbbreviatedCluster = f"CL{self.ClusterNumber}"
-        self.CallManagerGroup = "2A"
-        self.CAC = 1 
-        self.TZ = "CMLocal" # TODO: value checking
-        self.Carrier = "ATT" # TODO: value checking
-
-    def Build(self):
-        conn = AxlConnection(WSDL)
-        if conn.Open():
-            try:
-                for i in range(1,100):
-                    if i == 1:
-                        successful, data = BuildRegion(conn.Service, f"{self.SiteCode}_{self.AbbreviatedCluster}_R", self.AbbreviatedCluster)
-                    elif i == 2:
-                        BuildLocation(conn.Service, self.SiteCode, self.AbbreviatedCluster, self.CAC)
-                    elif i == 3:
-                        successful, data = BuildDevicePool(conn.Service, self.SiteCode, self.AbbreviatedCluster, self.CallManagerGroup, self.TZ, f"SBC_{self.AbbreviatedCluster}_{self.Carrier}_RG")
-                    else:
-                        break
-
-                    if not successful:
-                        raise Exception(data)
-
-                print("Site build completed successfully")
-                return True
-
-            except Exception as err:
-                strErr = f"{err}"
-                if strErr[0:24] == "Could not insert new row":
-                    print("Error: site already exists!")
-                else:
-                    print(strErr)
-                return False
-        else:
-            print("Error Opening AXL Connection")
-            return False
-
-
-
-#
-# Non-Instanced functions below
-#
-
-def BuildRegion(conn, RegionName, AbbrevCluster, AddRegionMatrices = True):
+def Region(conn, RegionName, AbbrevCluster, AddRegionMatrices = True):
     try:
         resp = conn.addRegion(region={"name" : f"{RegionName}"})
         region_uuid = resp["return"].strip("{}").lower()
@@ -133,7 +24,7 @@ def BuildRegion(conn, RegionName, AbbrevCluster, AddRegionMatrices = True):
 def __getRegionUUIDs(conn, AbbrevCluster):
     try:
         ListUUID = []
-        for region in RegionNamesList:
+        for region in Config.RegionNamesList:
             resp = conn.getRegion(name = f"{region}_{AbbrevCluster}_R")
             ListUUID.append(resp['return']['region']['uuid'].strip("{}").lower())
         return True, ListUUID
@@ -154,7 +45,7 @@ def addRegionMatrix(conn, Aregionuuid, Bregionuuid):
     except Fault as err:
         return False, err
 
-def BuildLocation(conn, SiteCode, AbbrevCluster, CAC, VideoBandwidth = 512, AssociateE911 = True):
+def Location(conn, SiteCode, AbbrevCluster, CAC, VideoBandwidth = 512, AssociateE911 = True):
     LocationDict = {
         'name' : f"{SiteCode}_{AbbrevCluster}_L",
         'withinAudioBandwidth' : 0,
@@ -191,7 +82,7 @@ def BuildLocation(conn, SiteCode, AbbrevCluster, CAC, VideoBandwidth = 512, Asso
     except Exception as err:
         return False, err
 
-def BuildDevicePool(conn, SiteCode, AbbrevCluster, CMRG, TZ, StandardLRG = None):
+def DevicePool(conn, SiteCode, AbbrevCluster, CMRG, TZ, StandardLRG = None):
     try: 
         resp = conn.addDevicePool(
             devicePool = {
