+++ RING/scripts/test.py
