+++ RING/scripts/test3.py
